python_for_devops/monitoring_and_logging/metrics-demo/requirements.txt
python-statsd
boto3
